Naprawa i refaktoryzacja testów (#46)
