+++ .github/workflows/tests.yml
@@ -20,5 +20,4 @@ jobs:
     - run: flutter pub get
     - run: flutter analyze
       continue-on-error: true
-    # Run unit tests
-    - run: flutter test test/unit_tests.dart
+    - run: flutter test
