+++ lib/pages/scan/scan_vibration.dart
@@ -1,0 +1,12 @@
+import 'package:vibration/vibration.dart';
+
+abstract class ScanVibration {
+  void vibrate();
+}
+
+class ScanVibrationImpl implements ScanVibration {
+  @override
+  void vibrate() {
+    Vibration.vibrate(duration: 200);
+  }
+}
